Backend/requirements.txt
# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=8.0.0
PyYAML>=5.3.1
scipy>=1.4.1
torch==1.7.1
torchvision==0.8.2
tqdm>=4.41.0
flask_restful   
Flask-API
Flask
flask_migrate
flask_sqlalchemy
flask_cors
PyJWT==2.1.0
bson
python-dotenv
dataclasses-json
websocket-client
APScheduler
websockets
easydict
psycopg2
configparser
psutil
# pytracemalloc
# logging -------------------------------------
tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas